whitespace fixes (now uses tabs) + little polish
People generally prefer to use tabs because they can set the width of
them in the editor. Myself, I usually use 2-spaces but I'll try to
remember to run spaces-to-tabs before I commit. Cuz I just loaded this
up in an IDE with 4-space tabs and it was a mess.